langchain-templates/langchain-starter-example/pyproject.toml
[project]
name = "steve-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.15",
    "langchain-chroma>=0.2.0",
    "langchain-community>=0.3.15",
    "langchain-openai>=0.3.2",
    "python-dotenv>=1.0.1",
    "pypdf>=5.1.0",
]